+++ docs/changelog.rst
@@ -47,6 +47,7 @@ Fixes
   for the planned case study experiment
 - Fixed the radicalisation checks in :meth:`~gatoh.model.model.ABModel.iteration_opinion_changes` to stop repeated deradicalisation of agents
 - **MAJOR:** Fixed numerous variable declarations and added manual garbage collection throughout various functions to prevent memory leaks during model runtimes
+- Fixed modulo operations in new :class:`~gatoh.visualisation.visualisation.ABVisualiser` methods.
 
 0.1.0-alpha
 ===========